image_quality_check/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
numpy==2.1.3
opencv-python-headless==4.12.0.88
Pillow==11.3.0
image-quality==1.2.7
python-multipart==0.0.6
python-dotenv==1.0.1
